test: add dedicated server and chromedriver hooks
